2.7.3：使用固定位数的整型代替原来的整型类型的别名声明

+++ src/Algorithms/Factorial.h
@@ -21,7 +21,7 @@ namespace CZ
 
     namespace Factorial_Impl
     {
-        using ValueType = long long;
+        using ValueType = int64_t;
 
         ValueType recursion_fact(const ValueType x)
         {
